backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
firebase-admin==6.5.0
google-generativeai==0.8.3
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.5.2
websockets==13.1
httpx==0.27.2
python-multipart==0.0.12
pytest==8.3.4
